+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/abc/avm2/parser/script/ActionScript3Parser.java
@@ -91,6 +91,7 @@ import com.jpexs.decompiler.flash.abc.avm2.model.operations.URShiftAVM2Item;
 import com.jpexs.decompiler.flash.abc.avm2.model.operations.UnPlusAVM2Item;
 import com.jpexs.decompiler.flash.abc.avm2.parser.AVM2ParseException;
 import com.jpexs.decompiler.flash.abc.types.Float4;
+import com.jpexs.decompiler.flash.abc.types.MethodBody;
 import com.jpexs.decompiler.flash.abc.types.Namespace;
 import com.jpexs.decompiler.flash.abc.types.ScriptInfo;
 import com.jpexs.decompiler.flash.action.swf4.ActionIf;
@@ -317,15 +318,22 @@ public class ActionScript3Parser {
                 s = lex();
             } else {
                 s = lex();
-                expected(s, lexer.yyline(), SymbolGroup.IDENTIFIER, SymbolType.MULTIPLY);
+                expected(s, lexer.yyline(), SymbolGroup.IDENTIFIER, SymbolType.MULTIPLY,
+                        SymbolType.PUBLIC, SymbolType.PROTECTED, SymbolType.PRIVATE, SymbolType.INTERNAL);
                 String propName = s.value.toString(); //Can be *
+                String nsKeyword = null;
+                if (s.isType(SymbolType.PUBLIC, SymbolType.PROTECTED, SymbolType.PRIVATE, SymbolType.INTERNAL)) {
+                    nsKeyword = s.value.toString();
+                }
                 GraphTargetItem propItem = null;
                 s = lex();
                 String nsSuffix = "";
                 GraphTargetItem ns = null;
                 if (s.type == SymbolType.NAMESPACE_OP) {
-                    ns = new UnresolvedAVM2Item(new ArrayList<>(), importedClasses, false, null, lexer.yyline(), new DottedChain(new String[]{propName}, new String[]{""} /*FIXME ???*/), null, openedNamespaces, abcIndex);
-                    variables.add((UnresolvedAVM2Item) ns);
+                    if (nsKeyword == null) {
+                        ns = new UnresolvedAVM2Item(new ArrayList<>(), importedClasses, false, null, lexer.yyline(), new DottedChain(new String[]{propName}, new String[]{""} /*FIXME ???*/), null, openedNamespaces, abcIndex);
+                        variables.add((UnresolvedAVM2Item) ns);
+                    }
                     s = lex();
                     if (s.type == SymbolType.BRACKET_OPEN) {
                         propItem = expression(allOpenedNamespaces, thisType, pkg, needsActivation, importedClasses, openedNamespaces, registerVars, inFunction, inMethod, true, variables, false, abc);
@@ -337,6 +345,9 @@ public class ActionScript3Parser {
                         propItem = null;
                     }
                 } else {
+                    if (nsKeyword != null) {
+                        throw new AVM2ParseException(nsKeyword + " not expected in this situation", lexer.yyline());
+                    }
                     if (s.type == SymbolType.NAMESPACESUFFIX) {
                         nsSuffix = "#" + s.value;
                     } else {
@@ -347,7 +358,7 @@ public class ActionScript3Parser {
                 if (ns != null) {
                     ret = new NamespacedAVM2Item(ns, propName, propItem, ret, attr, openedNamespaces, null);
                 } else {
-                    ret = new PropertyAVM2Item(ret, attr, propName, nsSuffix, abcIndex, openedNamespaces, new ArrayList<>(), nullDot);
+                    ret = new PropertyAVM2Item(ret, attr, propName, nsSuffix, abcIndex, openedNamespaces, new ArrayList<>(), nullDot, nsKeyword);
                 }
                 s = lex();
             }
@@ -369,18 +380,30 @@ public class ActionScript3Parser {
             attribute = true;
             s = lex();
         }
-        expected(s, lexer.yyline(), SymbolGroup.IDENTIFIER, SymbolType.THIS, SymbolType.SUPER, SymbolType.STRING_OP);
+        expected(s, lexer.yyline(), SymbolGroup.IDENTIFIER, SymbolType.THIS, SymbolType.SUPER, SymbolType.STRING_OP,
+                SymbolType.PUBLIC, SymbolType.PROTECTED, SymbolType.PRIVATE, SymbolType.INTERNAL);
+        
+        String nsKeyword = null;
+        if (s.isType(SymbolType.PUBLIC, SymbolType.PROTECTED, SymbolType.PRIVATE, SymbolType.INTERNAL)) {
+            nsKeyword = s.value.toString();
+        }
         name2 += s.value.toString();
         s = lex();
         boolean attrBracket = false;
         String nsSuffix = "";
         if (s.type == SymbolType.NAMESPACESUFFIX) {
+            if (nsKeyword != null) {
+                throw new AVM2ParseException(nsKeyword + " not expected in this situation", lexer.yyline());
+            }
             s = lex();
             nsSuffix = "#" + s.value;
         }
 
         name = name.add(attribute, name2, nsSuffix);
-        while (s.isType(SymbolType.DOT)) {
+        while (s.isType(SymbolType.DOT)) {        
+            if (nsKeyword != null) {
+                throw new AVM2ParseException(nsKeyword + " not expected in this situation", lexer.yyline());
+            }
             //name += s.value.toString(); //. or ::
             s = lex();
             name2 = "";
@@ -400,12 +423,21 @@ public class ActionScript3Parser {
                     continue;
                 }
             } else {
-                expected(s, lexer.yyline(), SymbolGroup.IDENTIFIER, SymbolType.NAMESPACE, SymbolType.MULTIPLY);
+                expected(
+                        s, lexer.yyline(), SymbolGroup.IDENTIFIER, SymbolType.NAMESPACE, SymbolType.MULTIPLY,
+                        SymbolType.PUBLIC, SymbolType.PROTECTED, SymbolType.PRIVATE, SymbolType.INTERNAL
+                );
                 name2 += s.value.toString();
+                if (s.isType(SymbolType.PUBLIC, SymbolType.PROTECTED, SymbolType.PRIVATE, SymbolType.INTERNAL)) {
+                    nsKeyword = s.value.toString();
+                }
             }
             s = lex();
             nsSuffix = "";
             if (s.type == SymbolType.NAMESPACESUFFIX) {
+                if (nsKeyword != null) {
+                    throw new AVM2ParseException(nsKeyword + " not expected in this situation", lexer.yyline());
+                }
                 nsSuffix = "#" + s.value;
                 s = lex();
             }
@@ -427,6 +459,8 @@ public class ActionScript3Parser {
             }
             name = name.getWithoutLast();
             s = lex();
+        } else if (nsKeyword != null) {
+            throw new AVM2ParseException(nsKeyword + " not expected in this situation", lexer.yyline());
         }
 
         GraphTargetItem ret = null;
@@ -437,16 +471,20 @@ public class ActionScript3Parser {
             ret = unr;
         }
         if (nsname != null) {
-            UnresolvedAVM2Item ns = new UnresolvedAVM2Item(new ArrayList<>(), importedClasses, typeOnly, null, lexer.yyline(), new DottedChain(new String[]{nsname}), null, openedNamespaces, abcIndex);
-            variables.add(ns);
-            ret = new NamespacedAVM2Item(ns, nsprop, nspropItem, ret, nsAtribute, openedNamespaces, null);
+            if (nsKeyword != null && nspropItem == null) {
+                ret = new PropertyAVM2Item(ret, nsAtribute, nsprop, nsSuffix, abcIndex, openedNamespaces, new ArrayList<MethodBody>(), false, nsKeyword);
+            } else {
+                UnresolvedAVM2Item ns = new UnresolvedAVM2Item(new ArrayList<>(), importedClasses, typeOnly, null, lexer.yyline(), new DottedChain(new String[]{nsname}), null, openedNamespaces, abcIndex);
+                variables.add(ns);
+                ret = new NamespacedAVM2Item(ns, nsprop, nspropItem, ret, nsAtribute, openedNamespaces, null);
+            }
         }
-        if (s.type == SymbolType.BRACKET_OPEN) {
+        if (s.isType(SymbolType.BRACKET_OPEN, SymbolType.DOT, SymbolType.NULL_DOT, SymbolType.TYPENAME)) {
             lexer.pushback(s);
-            if (attrBracket) {
+            /*if (attrBracket) {
                 lexer.pushback(new ParsedSymbol(-1, SymbolGroup.OPERATOR, SymbolType.ATTRIBUTE, "@"));
                 lexer.pushback(new ParsedSymbol(-1, SymbolGroup.OPERATOR, SymbolType.DOT, "."));
-            }
+            }*/
             ret = member(allOpenedNamespaces, thisType, pkg, needsActivation, importedClasses, openedNamespaces, ret, registerVars, inFunction, inMethod, variables, abc);
         } else {
             lexer.pushback(s);
@@ -586,7 +624,7 @@ public class ActionScript3Parser {
         if (hasRest) {
             subvariables.add(new NameAVM2Item(TypeItem.UNBOUNDED, lexer.yyline(), false, paramNames.get(paramNames.size() - 1), "", null, true, openedNamespaces, abcIndex, false));
         }
-        subvariables.add(new NameAVM2Item(thisType, lexer.yyline(), false, "arguments", "", null, true, openedNamespaces, abcIndex, false));
+        subvariables.add(new NameAVM2Item(TypeItem.ARRAY, lexer.yyline(), false, "arguments", "", null, true, openedNamespaces, abcIndex, false));
         int parCnt = subvariables.size();
         Reference<Boolean> needsActivation2 = new Reference<>(false);
         if (!isInterface && !isNative) {
@@ -648,7 +686,7 @@ public class ActionScript3Parser {
         lexer.pushback(s);
         return metadata;
     }
-
+    
     private void classTraits(List<List<NamespaceItem>> allOpenedNamespaces, boolean outsidePackage, List<AssignableAVM2Item> cinitVariables, Reference<Boolean> cinitNeedsActivation, List<GraphTargetItem> cinit, List<DottedChain> importedClasses, List<NamespaceItem> openedNamespaces, NamespaceItem pkg, String classNameStr, boolean isInterface, List<GraphTargetItem> traits, List<AssignableAVM2Item> iinitVariables, Reference<Boolean> iinitNeedsActivation, Reference<GraphTargetItem> iinit, ABC abc) throws AVM2ParseException, IOException, CompilationException, InterruptedException {
 
         NamespaceItem publicNs = new NamespaceItem("", Namespace.KIND_PACKAGE);
@@ -2739,6 +2777,10 @@ public class ActionScript3Parser {
             case THIS:
             case SUPER:
             case ATTRIBUTE:
+            case PUBLIC:
+            case PRIVATE:
+            case PROTECTED:
+            case INTERNAL:
                 lexer.pushback(s);
                 ret = name(allOpenedNamespaces, thisType, pkg, needsActivation, false, openedNamespaces, registerVars, inFunction, inMethod, variables, importedClasses, abc);
                 allowMemberOrCall = true;