+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/abc/avm2/parser/script/CallAVM2Item.java
@@ -139,7 +139,7 @@ public class CallAVM2Item extends AVM2Item {
                     nobj.setRegNumber(0);
                     obj = nobj;
                 }*/
-                PropertyAVM2Item p = new PropertyAVM2Item(obj, n.isAttribute(), n.getVariableName(), n.getNamespaceSuffix(), g.abcIndex, n.openedNamespaces, new ArrayList<>(), false);
+                PropertyAVM2Item p = new PropertyAVM2Item(obj, n.isAttribute(), n.getVariableName(), n.getNamespaceSuffix(), g.abcIndex, n.openedNamespaces, new ArrayList<>(), false, null);
                 p.setAssignedValue(n.getAssignedValue());
                 callable = p;
             }