+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 - AS3 property resolving for KIND_NAMESPACE (like builtin for Strings, etc.)
 - [#2636] ActionScript - switches vs loop breaks
 - [#2636] ActionScript 3 - type coercion / convert, local registers type propagation
+- Using public/private/protected/internal keywords on ambiguous namespace references
+  and compiling such keywords with `::` suffix.
 
 ## [25.1.0] - 2026-02-17
 ### Added